split store.go into specific files
